Create inits for Interpolator classes

+++ lib/ios/Interpolators/Interpolator.h
@@ -10,6 +10,11 @@
 
 @protocol Interpolator<NSObject>
 
-- (float) interpolate:(float)progress;
+/**
+ * Interpolates the progress value on a custom interpolation function such as Ease/Bezier.
+ * In Mathematical terms, the `progress` parameter is the `x` value of the interpolation function `f(x)`,
+ * while the return value is `f` (or `y`)
+ */
+- (CGFloat) interpolate:(CGFloat)progress;
 
 @end
